Fix homepage
The homepage has not been consistently falling back to a normal section
(list) layout, so explicity set the homepage layout.

We use a partial because we will be sharing the layout (just a list of
pages (regular and section) that are below the current section) with
the default section layout and making the current default section layout
(which is paginated) only used when set to a custom layout.

Signed-off-by: Daniel F. Dickinson <cshored@danielfdickinson.ca>