Use a simpler default section layout
We switch the default section layout to a simple list of pages (sections and
pages) in the section, using the same partial as for the current homepage
layout (to keep the default layout truly minimal).

We still allow testing of pagination through setting the 'layout'
metadata on a section page to 'paginated'.

Signed-off-by: Daniel F. Dickinson <cshored@danielfdickinson.ca>